fixed some compiler warnings related to signed integers; rework program to always use unsigned ints when working with scan numbers and other index variables moved definition of compile-time constants out of main and into the global_vars header

+++ include/qalgorithms_qpeaks.h
@@ -47,7 +47,7 @@ namespace qAlgorithms
         const float *y_start,
         const float *ylog_start,
         const bool *df_start,
-        const int n,
+        const size_t n,
         std::vector<RegressionGauss> &validRegressions);
 
     void validateRegressions(
